viddertech_quiz_bot/requirements.txt
python-telegram-bot[ext]
python-dotenv
openai
telethon